Introduce LoggableCollectionTrait
Since `GridFS` has `Collection` type, `LoggableGridFS` can share same
loggable methods as `LoggableCollection`. Therefore,
`LoggableCollectionTrait` having extracted methods from
`LoggableCollection` was introduced in this commit.
